Load country-wide data on monthly earnings by education

Loads data on monthly earnings by education level from one or more local Excel files downloaded from ISPV links retrieved by pv_list_reg().

Usage

pv_cr_monthlypay_education(path, sheet = 3)

Arguments

path

path(s) to file(s), Will be file with "CR_YYQD and either "PLS" or "MZS" in the name.

sheet

sheet number; you should be able to leave this as default (3) if using files downloaded from ISPV

Value

a tibble

Examples

pv_cr_monthlypay_education(system.file("extdata", "CR_204_PLS.xlsx", package = "ispv"))
#> # A tibble: 7 × 20
#>   category     code  fte_t…¹ pay_m…² pay_m…³ pay_d1 pay_q1 pay_q3 pay_d9 pay_m…⁴
#>   <chr>        <chr>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#> 1 Základní a … A-C      18.7  23677.    116. 16903. 19851. 30684. 37994.  25927.
#> 2 Střední bez… D,E,…    91.8  27805.    113. 20014. 22853. 33758. 39286.  29021.
#> 3 Střední s m… K-M     243.   37925.    109. 27294. 32277. 45802. 54941.  39854.
#> 4 Vyšší odbor… N,P,R    75.1  42642.    110. 30408. 35330. 52666. 62759.  44987.
#> 5 Vysokoškols… T,V     202.   47403.    109. 35854. 41317. 57955. 75660.  52872.
#> 6 Neuvedeno    NA       17.1  38828.    113. 26897. 33690. 44131. 51214.  39308.
#> 7 CELKEM - pl… NA      648.   39900.    110. 25228. 32199. 49028. 61312.  42555.
#> # … with 10 more variables: pay_mean_yoy <dbl>, bonus_perc <dbl>,
#> #   supplements_perc <dbl>, compensation_perc <dbl>, hours_per_month <dbl>,
#> #   file <chr>, sfera <chr>, period <chr>, year <chr>, quarter <chr>, and
#> #   abbreviated variable names ¹​fte_thous, ²​pay_median, ³​pay_median_yoy,
#> #   ⁴​pay_mean
pv_cr_monthlypay_education(system.file("extdata", "CR_204_MZS.xlsx", package = "ispv"))
#> # A tibble: 7 × 20
#>   category     code  fte_t…¹ pay_m…² pay_m…³ pay_d1 pay_q1 pay_q3 pay_d9 pay_m…⁴
#>   <chr>        <chr>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#> 1 Základní a … A-C     225.   25634.    109. 16096. 19692. 32007. 3.90e4  27178.
#> 2 Střední bez… D,E,…   998.   28180.    104. 17498. 21715. 35304. 4.32e4  29718.
#> 3 Střední s m… K-M    1011.   32740.    105. 19468. 25214. 42978. 5.68e4  36951.
#> 4 Vyšší odbor… N,P,R   151.   37326.    105. 21791. 28339. 51074. 7.32e4  44702.
#> 5 Vysokoškols… T,V     457.   47686.    102. 24563. 33616. 72093. 1.08e5  61016.
#> 6 Neuvedeno    NA       51.5  26368.    108. 16130. 19916. 34519. 4.61e4  30738.
#> 7 CELKEM - mz… NA     2893.   31567.    105. 18402. 23915. 42283. 6.01e4  37789.
#> # … with 10 more variables: pay_mean_yoy <dbl>, bonus_perc <dbl>,
#> #   supplements_perc <dbl>, compensation_perc <dbl>, hours_per_month <dbl>,
#> #   file <chr>, sfera <chr>, period <chr>, year <chr>, quarter <chr>, and
#> #   abbreviated variable names ¹​fte_thous, ²​pay_median, ³​pay_median_yoy,
#> #   ⁴​pay_mean