Load regional data on monthly earnings by profession

Load ISPV Excel files with regional data on monthly earnings by ISCO code

Usage

pv_reg_monthlypay_age_gender(path, sheet = 2)

Arguments

path

path(s) to file(s), Will be file with "Reg_YYQ" and either "PLS" or "MZS" in the name.

sheet

sheet number; you should be able to leave this as default (2) if using files downloaded from ISPV

Value

a tibble

Examples

pv_reg_monthlypay_age_gender(system.file("extdata", "Kar_204_pls.xlsx", package = "ispv"))
#> # A tibble: 21 × 23
#>    kraj_id…¹ categ…² fte_t…³ pay_m…⁴ pay_m…⁵ pay_d1 pay_q1 pay_q3 pay_d9 pay_m…⁶
#>    <chr>     <chr>     <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1 Kar       CELKEM… 15.1     38544.    109. 24578. 31164. 46771. 55221.  39930.
#>  2 Kar       do 20 …  0.0176     NA      NA     NA     NA     NA     NA      NA 
#>  3 Kar       20 - 2…  1.20    34570.    112. 24312. 28787. 39544. 43640.  34258.
#>  4 Kar       30 - 3…  2.49    37657.    108. 25261. 30450. 44988. 50339.  38043.
#>  5 Kar       40 - 4…  4.73    39082.    109. 25373. 31735. 47561. 56925.  40617.
#>  6 Kar       50 - 5…  4.82    39625.    109. 24113  32189. 48704. 58708.  41419.
#>  7 Kar       60 a v…  1.82    39167.    109. 23126. 31225. 48008. 55891.  40667.
#>  8 Kar       MUŽI     5.20    42524.    108. 26842. 34015. 50657. 60421.  43626.
#>  9 Kar       do 20 …  0.0086     NA      NA     NA     NA     NA     NA      NA 
#> 10 Kar       20 - 2…  0.525   38302.    111. 25344  31935. 42236. 44834.  36670.
#> # … with 11 more rows, 13 more variables: bonus_perc <dbl>,
#> #   supplements_perc <dbl>, compensation_perc <dbl>, hours_per_month <dbl>,
#> #   file <chr>, group <chr>, sfera <chr>, period <chr>, year <chr>,
#> #   quarter <chr>, kraj_id <chr>, kraj_name <chr>, kraj_id_nuts3 <chr>, and
#> #   abbreviated variable names ¹​kraj_id_ispv, ²​category, ³​fte_thous,
#> #   ⁴​pay_median, ⁵​pay_median_yoy, ⁶​pay_mean
pv_reg_monthlypay_age_gender(system.file("extdata", "Kar_204_mzs.xlsx", package = "ispv"))
#> # A tibble: 21 × 23
#>    kraj_id…¹ categ…² fte_t…³ pay_m…⁴ pay_m…⁵ pay_d1 pay_q1 pay_q3 pay_d9 pay_m…⁶
#>    <chr>     <chr>     <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1 Kar       CELKEM…  59.8    28898.    105. 18430. 22914. 36979. 47033.  32029.
#>  2 Kar       do 20 …   0.172  23380.    109. 17254. 20224. 27541. 31636.  24375.
#>  3 Kar       20 - 2…   7.87   28118.    106. 18073. 22914. 34173. 40702.  29343.
#>  4 Kar       30 - 3…  11.5    28154.    101. 17694. 22049. 36322. 47238.  31047.
#>  5 Kar       40 - 4…  18.3    29462.    103. 18925. 23410. 38589. 49593.  33272.
#>  6 Kar       50 - 5…  15.9    29559.    108. 18923. 23465. 37508. 47173.  32838.
#>  7 Kar       60 a v…   6.16   28233.    110. 18243. 22472. 35819. 45628.  31727.
#>  8 Kar       MUŽI     34.9    30614.    102. 19189. 24082. 38589. 48620.  33934.
#>  9 Kar       do 20 …   0.105  24987.    114. 17811. 21450. 30010. 31636.  25361.
#> 10 Kar       20 - 2…   4.79   29328.    104. 20255. 23726. 34611. 41692.  30378.
#> # … with 11 more rows, 13 more variables: bonus_perc <dbl>,
#> #   supplements_perc <dbl>, compensation_perc <dbl>, hours_per_month <dbl>,
#> #   file <chr>, group <chr>, sfera <chr>, period <chr>, year <chr>,
#> #   quarter <chr>, kraj_id <chr>, kraj_name <chr>, kraj_id_nuts3 <chr>, and
#> #   abbreviated variable names ¹​kraj_id_ispv, ²​category, ³​fte_thous,
#> #   ⁴​pay_median, ⁵​pay_median_yoy, ⁶​pay_mean