Load regional data on monthly earnings by education

Loads data on monthly earnings by ISCO code from one or more local Excel files downloaded from ISPV links retrieved by pv_list_reg().

Usage

pv_reg_monthlypay_education(path, sheet = 2)

Arguments

path

path(s) to file(s), Will be file with "Reg_YYQ" and either "PLS" or "MZS" in the name.

sheet

sheet number; you should be able to leave this as default (2) if using files downloaded from ISPV

Value

a tibble

Examples

pv_reg_monthlypay_education(system.file("extdata", "Kar_204_pls.xlsx", package = "ispv"))
#> # A tibble: 7 × 23
#>   kraj_id_ispv categ…¹ code  fte_t…² pay_m…³ pay_m…⁴ pay_d1 pay_q1 pay_q3 pay_d9
#>   <chr>        <chr>   <chr>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1 Kar          Základ… A-C     0.515  21674.    113. 15715. 18645. 27292. 33804.
#> 2 Kar          Středn… D,E,…   1.86   26882.    114. 19530. 22247. 32290. 38086.
#> 3 Kar          Středn… K-M     6.77   36790.    108. 26773. 31295. 43586. 50745.
#> 4 Kar          Vyšší … N,R     1.59   41519.    108. 30604. 34801. 51020. 60552.
#> 5 Kar          Vysoko… T,V     3.70   47039.    111. 36226. 41769. 53262. 67570.
#> 6 Kar          Neuved… NA      0.642  39549.    113. 31699. 35745. 43707. 49163.
#> 7 Kar          CELKEM… NA     15.1    38544.    109. 24578. 31164. 46771. 55221.
#> # … with 13 more variables: pay_mean <dbl>, bonus_perc <dbl>,
#> #   supplements_perc <dbl>, compensation_perc <dbl>, hours_per_month <dbl>,
#> #   file <chr>, sfera <chr>, period <chr>, year <chr>, quarter <chr>,
#> #   kraj_id <chr>, kraj_name <chr>, kraj_id_nuts3 <chr>, and abbreviated
#> #   variable names ¹​category, ²​fte_thous, ³​pay_median, ⁴​pay_median_yoy
pv_reg_monthlypay_education(system.file("extdata", "Kar_204_mzs.xlsx", package = "ispv"))
#> # A tibble: 7 × 23
#>   kraj_id_ispv categ…¹ code  fte_t…² pay_m…³ pay_m…⁴ pay_d1 pay_q1 pay_q3 pay_d9
#>   <chr>        <chr>   <chr>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1 Kar          Základ… A-C      8.28  24696.    107. 16661. 20221. 30243. 36310.
#> 2 Kar          Středn… D,E,…   24.0   26814.    101. 17516. 22160. 34339. 41206.
#> 3 Kar          Středn… K-M     18.8   30986.    105. 20008. 24569. 40073. 49988.
#> 4 Kar          Vyšší … N,P,R    1.84  34849.    103. 22663. 24777. 45374. 54374.
#> 5 Kar          Vysoko… T,V      3.50  42214.    113. 20463. 28177. 65381. 97233.
#> 6 Kar          Neuved… NA       3.32  30400.    120. 20496. 26176. 38473. 42299.
#> 7 Kar          CELKEM… NA      59.8   28898.    105. 18430. 22914. 36979. 47033.
#> # … with 13 more variables: pay_mean <dbl>, bonus_perc <dbl>,
#> #   supplements_perc <dbl>, compensation_perc <dbl>, hours_per_month <dbl>,
#> #   file <chr>, sfera <chr>, period <chr>, year <chr>, quarter <chr>,
#> #   kraj_id <chr>, kraj_name <chr>, kraj_id_nuts3 <chr>, and abbreviated
#> #   variable names ¹​category, ²​fte_thous, ³​pay_median, ⁴​pay_median_yoy