Load regional data on monthly earnings by profession

Loads data on monthly earnings by ISCO code from one or more local Excel files downloaded from ISPV links retrieved by pv_list_reg().

Usage

pv_reg_monthlypay_isco4(path, sheet = 4)

Format

The function returns a data frame with 22 variables. All earnings are monthly gross in CZK.

kraj_id_ispv

character. Internal region code, appears in Excel file name

file

character. File from which this row was read.

isco4_full

character. 4-digit ISCO (occupation) code with Czech ISCO category name. CZSO codelist number 5708

fte_thous

double. Thousand FTEs in this ISCO-4 category - converted to FTE by months paid.

pay_median

double. Median monthly earnings.

pay_d1

double. 1st decile monthly earnings.

pay_q1

double. 1st quartile monthly earnings.

pay_q3

double. 3rd quartile monthly earnings.

pay_d9

double. 9th decline monthly earnings.

pay_mean

double. Mean earnings.

bonus_perc

double. Bonuses ("odměny") as share of pay, as decimal

supplements_perc

double. Supplements ("příplatky") as share of pay, as decimal.

compensation_perc

double. Compensation ("náhrady") as share of pay, as decimal.

hours_per_month

double. Monthly hours worked.

sfera

character. Sphere - salary (pls) or wage (mzs), roughly equals public or private sector

period

character. Time period as appears in file name, e.g. 204 is Q4 of 2020. Regional data only comes in Q4, i.e. for full year.

year

character. Year, as 4-digit character vector

isco4_id

character. 4-digit ISCO code.

isco4_name

character. Czech name of 4-digit ISCO category

kraj_id

character. non-NUTS ID of kraj (region, NUTS3).

kraj_name

character. Czech name of kraj (region, NUTS3).

kraj_id_nuts3

character. NUTS ID of kraj (region, NUTS3)..

Arguments

path

path(s) to file(s), Will be file with "Reg_YYQ" and either "PLS" or "MZS" in the name.

sheet

sheet number; you should be able to leave this as default (4) if using files downloaded from ISPV

Value

a tibble, see Format for details.

Examples

pv_reg_monthlypay_isco4(system.file("extdata", "Kar_204_mzs.xlsx", package = "ispv"))
#> # A tibble: 81 × 23
#>    kraj_id…¹ isco4…² fte_t…³ pay_m…⁴ pay_d1 pay_q1 pay_q3 pay_d9 pay_m…⁵ bonus…⁶
#>    <chr>     <chr>     <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>
#>  1 Kar       1211 Ř…  0.091   63797. 44823. 50958. 9.37e4 1.37e5  77526.   15.3 
#>  2 Kar       1321 Ř…  0.205   82598. 38027. 49981. 9.78e4 1.25e5  82159.   17.1 
#>  3 Kar       1324 Ř…  0.106   52086. 34483. 41257. 6.58e4 8.56e4  56839.   19.2 
#>  4 Kar       1342 Ř…  0.0797  76681. 42009. 52869. 1.08e5 1.72e5  88861.    7.08
#>  5 Kar       1346 Ř…  0.0448  69291. 44588. 61697. 8.33e4 1.27e5  83029.   27.8 
#>  6 Kar       1420 Ř…  0.288   43681. 28790. 36769. 6.61e4 1.38e5  58149.   14.6 
#>  7 Kar       2132 S…  0.0931  44842. 33570. 39084. 4.71e4 5.09e4  43562.    9.87
#>  8 Kar       2141 S…  0.145   50817. 37180. 43468. 7.10e4 8.47e4  59174.   12.4 
#>  9 Kar       2144 S…  0.121   54324. 39622. 44146. 6.38e4 7.05e4  55599.   10.5 
#> 10 Kar       2145 C…  0.038   56774. 36814. 45886. 7.49e4 9.23e4  60815.    9.49
#> # … with 71 more rows, 13 more variables: supplements_perc <dbl>,
#> #   compensation_perc <dbl>, hours_per_month <dbl>, file <chr>, sfera <chr>,
#> #   period <chr>, year <chr>, quarter <chr>, kraj_id <chr>, kraj_name <chr>,
#> #   kraj_id_nuts3 <chr>, isco4_id <chr>, isco4_name <chr>, and abbreviated
#> #   variable names ¹​kraj_id_ispv, ²​isco4_full, ³​fte_thous, ⁴​pay_median,
#> #   ⁵​pay_mean, ⁶​bonus_perc
pv_reg_monthlypay_isco4(system.file("extdata", "Kar_204_pls.xlsx", package = "ispv"))
#> # A tibble: 52 × 23
#>    kraj_id…¹ isco4…² fte_t…³ pay_m…⁴ pay_d1 pay_q1 pay_q3 pay_d9 pay_m…⁵ bonus…⁶
#>    <chr>     <chr>     <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>
#>  1 Kar       1112 N…  0.0423  68663. 56052. 64480. 85739. 1.03e5  76977.   15.0 
#>  2 Kar       1211 Ř…  0.0419  56678. 42303. 46987. 68375. 8.59e4  60849.   11.2 
#>  3 Kar       1212 Ř…  0.0325  48122. 41411. 44095. 63436. 7.42e4  53019.   13.4 
#>  4 Kar       1219 O…  0.0714  56283. 39931. 44905. 73506. 8.93e4  60966.   15.1 
#>  5 Kar       1341 Ř…  0.125   54150. 44734. 48678. 58765. 6.32e4  54071.   12.4 
#>  6 Kar       1345 Ř…  0.255   67589. 51219. 59905. 74106. 7.97e4  66382.   14.0 
#>  7 Kar       1349 Ř…  0.0763  64568. 48317. 57700. 70540. 8.77e4  66798.    9.14
#>  8 Kar       2221 V…  0.134   61774. 44912. 49002. 69907. 7.55e4  60004.   12.4 
#>  9 Kar       2320 U…  0.301   43450. 37464. 39377. 48715. 5.20e4  44437.   10.6 
#> 10 Kar       2330 U…  1.09    45488. 38434. 41694. 49006. 5.21e4  45640.   13.2 
#> # … with 42 more rows, 13 more variables: supplements_perc <dbl>,
#> #   compensation_perc <dbl>, hours_per_month <dbl>, file <chr>, sfera <chr>,
#> #   period <chr>, year <chr>, quarter <chr>, kraj_id <chr>, kraj_name <chr>,
#> #   kraj_id_nuts3 <chr>, isco4_id <chr>, isco4_name <chr>, and abbreviated
#> #   variable names ¹​kraj_id_ispv, ²​isco4_full, ³​fte_thous, ⁴​pay_median,
#> #   ⁵​pay_mean, ⁶​bonus_perc