Load ISPV Excel files with regional data on monthly earnings by ISCO code

Loads data on monthly earnings by ISCO code from one or more local Excel files downloaded from ISPV links retrieved by pv_list_reg().

Usage

pv_reg_monthlypay_isco4(path)

Format

The function returns a data frame with 22 variables. All earnings are monthly gross in CZK.

kraj_id_ispv

character. Internal region code, appears in Excel file name

file

character. File from which this row was read.

isco4_full

character. 4-digit ISCO (occupation) code with Czech ISCO category name. CZSO codelist number 5708

fte_thous

double. Thousand FTEs in this ISCO-4 category - converted to FTE by months paid.

pay_median

double. Median monthly earnings.

pay_d1

double. 1st decile monthly earnings.

pay_q1

double. 1st quartile monthly earnings.

pay_q3

double. 3rd quartile monthly earnings.

pay_d9

double. 9th decline monthly earnings.

pay_mean

double. Mean earnings.

bonus_perc

double. Bonuses ("odměny") as share of pay, as decimal

supplements_perc

double. Supplements ("příplatky") as share of pay, as decimal.

compensation_perc

double. Compensation ("náhrady") as share of pay, as decimal.

hours_per_month

double. Monthly hours worked.

sfera

character. Sphere - salary (pls) or wage (mzs), roughly equals public or private sector

period

character. Time period as appears in file name, e.g. 204 is Q4 of 2020. Regional data only comes in Q4, i.e. for full year.

year

character. Year, as 4-digit character vector

isco4_id

character. 4-digit ISCO code.

isco4_name

character. Czech name of 4-digit ISCO category

kraj_id

character. non-NUTS ID of kraj (region, NUTS3).

kraj_name

character. Czech name of kraj (region, NUTS3).

kraj_id_nuts3

character. NUTS ID of kraj (region, NUTS3)..

Arguments

path

path to one or more Excel files (character vector length one or more).

Value

a tibble, see Format for details.

Examples

pv_reg_monthlypay_isco4(system.file("extdata", "Kar_204_pls.xlsx", package = "ispv"))
#> # A tibble: 52 × 22
#>    kraj_id_ispv isco4_full      fte_thous pay_median pay_d1 pay_q1 pay_q3 pay_d9
#>    <chr>        <chr>               <dbl>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 Kar          1112 Nejvyšší …    0.0423     68663. 56052. 64480. 85739. 1.03e5
#>  2 Kar          1211 Řídící pr…    0.0419     56678. 42303. 46987. 68375. 8.59e4
#>  3 Kar          1212 Řídící pr…    0.0325     48122. 41411. 44095. 63436. 7.42e4
#>  4 Kar          1219 Ostatní ř…    0.0714     56283. 39931. 44905. 73506. 8.93e4
#>  5 Kar          1341 Řídící pr…    0.125      54150. 44734. 48678. 58765. 6.32e4
#>  6 Kar          1345 Řídící pr…    0.255      67589. 51219. 59905. 74106. 7.97e4
#>  7 Kar          1349 Řídící pr…    0.0763     64568. 48317. 57700. 70540. 8.77e4
#>  8 Kar          2221 Všeobecné…    0.134      61774. 44912. 49002. 69907. 7.55e4
#>  9 Kar          2320 Učitelé o…    0.301      43450. 37464. 39377. 48715. 5.20e4
#> 10 Kar          2330 Učitelé S…    1.09       45488. 38434. 41694. 49006. 5.21e4
#> # … with 42 more rows, and 14 more variables: pay_mean <dbl>, bonus_perc <dbl>,
#> #   supplements_perc <dbl>, compensation_perc <dbl>, hours_per_month <dbl>,
#> #   file <chr>, sfera <chr>, period <chr>, year <chr>, kraj_id <chr>,
#> #   kraj_name <chr>, kraj_id_nuts3 <chr>, isco4_id <chr>, isco4_name <chr>