Load ISPV Excel files with regional data on monthly earnings by ISCO code

Loads data on monthly earnings by ISCO code from one or more local Excel files downloaded from ISPV links retrieved by pv_list_reg().

Usage

pv_reg_monthlypay_isco4(path)

Format

The function returns a data frame with 22 variables. All earnings are monthly gross in CZK.

kraj_id_ispv

character. Internal region code, appears in Excel file name

file

character. File from which this row was read.

isco4_full

character. 4-digit ISCO (occupation) code with Czech ISCO category name. CZSO codelist number 5708

fte_thous

double. Thousand FTEs in this ISCO-4 category - converted to FTE by months paid.

pay_median

double. Median monthly earnings.

pay_d1

double. 1st decile monthly earnings.

pay_q1

double. 1st quartile monthly earnings.

pay_q3

double. 3rd quartile monthly earnings.

pay_d9

double. 9th decline monthly earnings.

pay_mean

double. Mean earnings.

bonus_perc

double. Bonuses ("odměny") as share of pay, as decimal

supplements_perc

double. Supplements ("příplatky") as share of pay, as decimal.

compensation_perc

double. Compensation ("náhrady") as share of pay, as decimal.

hours_per_month

double. Monthly hours worked.

sfera

character. Sphere - salary (pls) or wage (mzs), roughly equals public or private sector

period

character. Time period as appears in file name, e.g. 204 is Q4 of 2020. Regional data only comes in Q4, i.e. for full year.

year

character. Year, as 4-digit character vector

isco4_id

character. 4-digit ISCO code.

isco4_name

character. Czech name of 4-digit ISCO category

kraj_id

character. non-NUTS ID of kraj (region, NUTS3).

kraj_name

character. Czech name of kraj (region, NUTS3).

kraj_id_nuts3

character. NUTS ID of kraj (region, NUTS3)..

Arguments

path

path to one or more Excel files (character vector length one or more).

Value

a tibble, see Format for details.

Examples

pv_reg_monthlypay_isco4(system.file("extdata", "Kar_204_pls.xlsx", package = "ispv"))
#> Warning: TRE pattern compilation error 'Invalid regexp'
#> Error in regexpr("(?<=\\_)2[0-9][2,4](?=\\_) ", ispv_all$file): invalid regular expression '(?<=\_)2[0-9][2,4](?=\_) ', reason 'Invalid regexp'